backend/requirements.txt
# Core
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
python-multipart==0.0.9

# Database
sqlalchemy==2.0.35
aiosqlite==0.20.0

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# AI / LLM
groq==0.11.0
langchain==0.3.0
langchain-community==0.3.0
langchain-huggingface==0.1.0
faiss-cpu==1.8.0
sentence-transformers==3.1.0

# PDF
reportlab==4.2.2

# HTTP
httpx==0.27.2

# WebSocket
websockets==12.0

# Utilities
pydantic==2.9.2
pydantic-settings==2.5.2
uuid6==2024.7.10

# ML / Risk Scorer
lightgbm==4.5.0
scikit-learn==1.5.2
pandas==2.2.3
numpy==1.26.4
joblib==1.4.2
